Add ability to specify subprojects (#4088)
### Problem

Including a subproject (with its own build root and own build-root-relative BUILD targets) within a project and depending on targets of the subproject was previously unsupported. All subproject BUILD targets would not resolve properly since they would be relative to the subproject's buildroot, not the main project's build root. 

Further discussion found in issue https://github.com/pantsbuild/pants/issues/3978

### Solution

Add a new configuration option, `subproject-roots`, that allows the user to specify roots of all subprojects. Upon resolving targets in a BUILD file, the BUILD file path is compared to subproject roots. If the BUILD file resides in a subproject, then the target is corrected (The relative path between the project root and the subproject root is prepended to the spec).

### Result

It is now possible to include subprojects in larger projects. This PR is only limited in scope; however. The subproject's targets are still built using the main project's configurations, etc. In the future, it may be a nicety to allow the subproject's targets to be built with the subproject's `pants` and `pants.ini` files.