[jvm-compile] Copy compile classpath into runtime classpath even if already defined
In #4309, the problem was caused by jvm_compile assuming it is always first in the compile order. This ensures the compile_classpath is mixed into the runtime classpath even if it has already been initialized.

Another approach would be to move the NodeBuild task to later in the compile goal, and make it an error for jvm_compile to receive a runtime classpath.