Binary builder task for the new python pipeline. (#4313)
The PEX_PATH merging mechanism only works when executing a pex. 
It doesn't allow you to literally merge multiple pexes into a single .pex file. 
Therefore the most straightforward way of building that .pex is to dump 
the sources and requirements into it from scratch, ignoring the partial 
pexes that are the products of GatherSources and ResolveRequirements.

To make this easy, this change breaks out the relevant pex-building code 
from those tasks, so it can be re-used in this new task.

The remainder of the logic was copied from the old task, as was the test.