[jvmdoc-gen] exclude internal targets not matching lang predicate
javadoc sometimes will fail to compile if the classpath contains classfiles for classes with the same packages as the sources being documented when those classfiles are created by scalac.

This patch removes non-jar libraries from the doc classpath if they don't match the language predicate. This should fix that error. See #4306