Fix reliance on symlinks in testdata.
The new python pipeline runs tests out of a chroot, as it should.
However symlinks are replaced by their contents when copied into
chroots (as they presumably should be), so tests that rely on
symlinks in testdata will now fail.

This change fixes one such test by creating a symlink-preserving
tarball out of the previously loose testdata, and untarring it in the test.