Misc fixes to python tasks: (#4323)
PythonISortTask shouldn't be a PythonTask, as it uses none
of the chroot-building facilities of that base class.

Fix a missing subsystem dep.

Remove an old template that is now consumed from another
location.

Add the --fast option to the new PytestRun, even though it
does nothing, because some scripts still reference it, and
having it here, even as a no-op, will streamline the switch
to the new pipeline in our repo.

Move PythonSetup/PythonRepos subsystems out of the python backend.
This is a longstanding TODO: Pants needs these for plugin loading even
if the repo contains no python code at all. So these belong in core
pants, not in the python backend. Note that these are not public API, and
their options scopes haven't changed, so this move does not require deprecation.