Simplify PythonSetup usage.
Separates it from PythonRepos and moves it back into the
python backend (but under the subsystems dir).

To do so, removes plugin_resover.py's usage of PythonSetup.
The only use was for the resolver_cache_ttl option, and that
option shouldn't really exist anyway (see below).

Deprecates the --resolver-cache-ttl option.
There's no benefit to allowing re-resolution of requirements.
If you're not content with "some version that satisfies a
requirement, no matter how old" then you should tighten up your
requirements... Re-resolving arbitrarily can only add to
confusion and lack of repeatability, so now we simply hard-code
a very large TTL (10 years) in the two places we need it (plus in
python_chroot.py, but that will go away once we're on the new pipeine)
and deprecate the option.

Centralizes all handling of interpreter selection constraints, which
were previously spread across three options in three different scopes.
This change moves the --constraints option from the new pipeline's
interpreter selection task into PythonSetup, and deprecates the other
two options.

Note that PythonRepos stays outside the python backend, but now
in its own file, as it is genuinely needed by the plugin_resolver.