[engine] Split SelectDependencies into SelectDependencies and SelectTransitive (#4334)
### Problem

This is a refactoring change with no functionalities added. it will simplify #4283 so we could optimize just for `SelectTransitive`.

### Solution

Introduce `SelectTransitive` with identical attributes with `SelectDependencies` (for now), and split the `if` `else` switch in node handling into each node. With that, `transitive` flag is not needed.

### Result

Test all passes.