[engine] Improve memory layout for Graph (#4333)
### Problem

Profiling has shown that about 1/3 of the total memory usage in the new engine is in the rust `Graph` struct (specifically, in the adjacency list HashSets and the HashMaps that map from Node->Id and Id->Entry.)

### Solution

First, removed the `cyclic_dependencies` adjacency list in favor of privately wrapping/labeling `Node` instances as Cyclic. This removed a number of special cases for cycles, and allowed for...

Second, introduced a dependency on [petgraph](https://docs.rs/petgraph/0.4.3/petgraph/index.html), which implements a compact/stable `Vec`-indexed Graph representation. This: 1) reduces the total number of allocations, and 2) avoids fragmentation in both partially-filled collections, and padded allocations.

### Result

CPU time is reduced by about 5%, and memory usage by about 30% (6.6GB -> 4.5GB). Fixes #4305.