[pantsd] Add an option to configure the watchman startup timeout. (#4332)
Fixes #4224

Problem

In large repos, the initial watch-project command can take longer than the default timeout of 5 seconds due to watchman startup. This can cause a SocketTimeout exception, which leads to a FSEventService crash and subsequently tears down the daemon.

Solution

Add an option to set the initial watchman client timeout, which is then set to the default timeout after the initial watch-project command is issued.

Result

After testing in our monorepo, I'm unable to reproduce the startup crash seen in #4224 with this option set to 30 seconds.