Fix product mapping for ivy resolve when libraries are not jar files (#4339)
### Problem

ivy resolution does not add non-jar (zip) artifacts to the JarImportProducts.  It defaults the file extension to `.jar` in one spot.

I noticed this when I was trying to use `unpack-jars` with a zip file.  When I do this, nothing is unzipped in .pants.d, but a similar BUILD with a jar file works as expected.

```
jar_library(name='groovy', jars=[ jar('org.codehaus.groovy', 'groovy-binary', '2.2.2', intransitive=True, ext='zip') ])
unpacked_jars(name="unpack_groovy", libraries=[':groovy'])
```

`./pants -ldebug unpack-jars :unpack_groovy` shows no output that indicates the zip is unpacked.  With the fix, this works.

### Solution

The code was carrying through most properties of the jar (name, org, rev), but not extension.  This caused the extension to default to `.jar`.  The fix carries through the extension.