implement exports literal in jvm_target (#4329)
### Problem (#4315)

Given a scenario with 3 jvm targets A, B, C.
A depends on B. B depends on C.
In some cases (annotations, inheritance, etc), even though A does not directly use C, C has to be in A's classpath for A to compile. With strict_deps turned on, this means C has to be in A's direct dependency list.

This causes trouble if B is a core library that thousands of targets depends on. Anytime B adds a new dependency, if strict_deps is True, then it is very likely that all those thousands of targets that depend on B have to include this new dependency in their BUILD files as well. Given that strict_deps is the direction we are heading, we have to find a way for libraries to "export" some of its dependencies.

### Solution

Implemented bazel-like "exports" literal (https://bazel.build/versions/master/docs/be/java.html#java_library.exports) in pants. Now pants user can specify which of a target's dependencies will be exported so the consumers would not need to specify/know about these in case of strict_deps.

An example of this usage would be:

```
java_library(
  name='A',
  dependencies=[
    'src/java/dep/it:needs',
  ]
  exports=[
    'src/java/dep/it:needs',
  ]
)
```

With strict_deps = True, any targets that depend on A will have access to 'src/java/dep/it:needs' as well (in compile scope, that means A will have 'src/java/dep/it:needs' in its classpath).