Better workaround for Shared race (#4348)
### Problem

In order to work around #4298, I pushed execution of every `Node` onto a `CpuPool` (meaning, in the worst case, `N` extra context switches for `N` nodes). But while investigating fixing that issue, I determined that there is a much cheaper workaround available (which probably also points to a reproduction).

### Solution

Removing the context switching into the `CpuPool` causes #4298 to reproduce very close to the roots of the graph (one of two roots completes, but the task that should be completing the second root never wakes up to execute it).

But, removing the `join_all` at the root and `wait`ing on the two roots independently works around the issue, _without_ the extra context switches.

### Result

The better workaround for #4298 improves the runtime for `./pants --enable-v2-engine list :: | wc -l` in the `pantsbuild/pants` repo from ~`7.8` seconds to ~`5.8` seconds: a 25% improvement.