Move all logic out of Context (#4343)
### Problem

The `Context` object used to be helpful because it held a reference to the `Externs` struct, but now that externs are static it is an unnecessary layer of indirection, and a catchall for stuff that is only used by one caller.

Additionally, we lost our useful cycle-detected message during #4333 due to a test without a good description.

### Solution

Inline or move all `Context` methods: some to `externs`, and some to callers. In particular, methods related to `Snapshot` storage and lifting were moved to that trait. And finally, restore our useful cycle message by converting `Noop` to an enum ordered by how relevant the message is likely to be, and then always propagating the highest priority message.

### Result

Less obfuscation around context, and better error messages.