Speed up typical go builds.
Resolving remote Go libraries involves injecting targets representing
the remote sources into the build graph, so they can be compiled locally.
This must happen for every target, valid or not, in every build.

Previously, however, this involved two time-consuming operations
per item in the transitive closure of all remote deps:

- A network call to fetch the go meta tag, to figure out the
  remote library root from the import path (at least for remote libraries
  fetched by cloning).

- Shelling out to `go list` to get a library's imports.

This change caches both results in workdir files, for reuse in
subsequent pants runs.  This restores some of the snappiness of
Go when run in Pants.

Note that the first of these two items is unversioned (we're essentially
caching the resullts of network calls that are not qualified by version).
The second, however, is versioned, as expected.