[engine] short circuit native engine build failures (#4353)
### Problem:
 When the native engine fails to compile, pants will still try to run. If pants is running with the engine enabled, this blows up and pushes the compiler errors out of the first screen in the terminal.

### Solution: 
If the binary was not built, stop bootstrapping. This adds a check to see if there is a binary created by building the native code. If not, bootstrap stops.

### Discussion
I tried using `set -e`, but that didn't work as well as I would like. First, `build_native_code` already dies on failure. We just silently ignore it. `set -e` revealed that because instead of continuing execution, the copy of the binary would fail. Relying on the copy failing felt a bit haphazard, so I put in an explicit check.