Improve our use of gofmt.
Previously we integrated as a checkstyle-type check, that only logged
errors but didn't correct them.  In fact, it didn't even do that, it
just logged the names of files that had style errors (and even that
in a convoluted way).

Now the checkstyle task actually reports the errors it found.

We also add a fmt task that actually fixes the errors.