[engine] Init Tasks before Scheduler (#4381)
### Problem

As described in #4375, the lifecycle of the `Tasks` and `Scheduler` are currently such that the `Tasks` need to remain mutable in the `Scheduler` after it has been created. This made expanding usage of the `Tasks` struct in #4371 (to use a `RuleGraph` to launch tasks) more difficult.

### Solution

Move the initialization of the `Tasks` struct before the initialization of the `Scheduler`, and then pass the `Tasks` into the `Scheduler`. Also, removed the unnecessary `RawScheduler` wrapper.

### Result

Fixes #4375