python-binary-create task maps all product directories to the same target (#4390)
Problem

This fixes #4389

Solution

I've simply noticed that the binary variable was erroneously used outside the scope of its for loop, and replaced that use with vt.targets, which was likely what the original author intended. Using the wrong variable, in this case, caused all pex mappings to be mapped from the same target (whatever wound up being the terminal for loop case).

Result

When this patch is applied, product mappings are assembled correctly. From the example in #4389 , here's what the product mapping looks like now:

ProductMapping(pex_archives) {
  PythonBinary(BuildFileAddress(BuildFile(docker/monitoring_sidecar/BUILD, FileSystemProjectTree(/Users/kyle.derr/Repositories/allclearid/odin/mono)), monitoring_sidecar)) => /Users/kyle.derr/Repositories/allclearid/odin/mono/.pants.d/binary/python-binary-create/0550b754d96b/docker.monitoring_sidecar.monitoring_sidecar/current
    [u'monitoring_sidecar.pex']
  PythonBinary(BuildFileAddress(BuildFile(tools/templating/BUILD, FileSystemProjectTree(/Users/kyle.derr/Repositories/allclearid/odin/mono)), configure)) => /Users/kyle.derr/Repositories/allclearid/odin/mono/.pants.d/binary/python-binary-create/0550b754d96b/tools.templating.configure/current
    [u'configure.pex']
}