[engine] Rules as decorators (#4369)
### Problem

The "rule triple" / "task triple" syntax generally requires ~30 minutes worth of explanation for every new user. In particular, the triples _look_ like they are decoupled from individual functions, when they are generally defined 1-to-1.

### Solution

Switch the majority of rule declarations to a new `@rule` decorator which creates and annotates the function with a `TaskRule`. The decorator is usable in all cases where constraints are known statically... in about three cases (where constraints are computed dynamically: ie, types from the `SymbolTable`) it is necessary to explicitly create a `TaskRule`.

Additionally, removed `SelectLiteral` in favor of usage of `SingletonRule`s in all cases. When a `SingletonRule` is declared, callers can directly use `Select` to request the value for the singleton. Just like `SelectLiteral`, these continue to be an escape hatch for values which are not constructed from the ground up using files on disk (see #3941).

### Result

Fixes #4289; "task triples" are gone: long live the `@rule` decorator!