Switch the new PytestRun task to use junitxml output.
This addresses https://github.com/pantsbuild/pants/issues/3837

This change also created a unit test for the new task, based
on the old task's tests with appropriate changes.  Somehow this
test wasn't created when the task was.

A few small tweaks were necessary to get the tests to pass.

This change also moves the existing pytest run integration test
into the directory containing all the other new-task tests, so
we don't forget it when we delete the old tasks and their tests.
That integration test has been running the new task since we
switched that to be the default in the repo anyway.

Note that the old task no longer has an integration test (and
hasn't since we made the switch to the new one in our repo)
but that seems fine since it's going to be deprecated and
removed soon.