A couple of test fixes.
In particular, note that in new pytest-cov there is no 'clever'
detection of files vs. modules.  Instead coverage just checks
if a specified source is an existing dir on the filesystem.
The upshot of this is that you can only specify dirs (relative
to the cwd) or dotted modules, but not individual .py files.