Fully hydrate a BuildGraph for the purposes of ChangedCalculator. (#4424)
### Problem

Unfortunately, using only `HydratedTarget` objects to walk the graph constructed by the engine will not include `traversable_dependency_specs` from `Target` objects that are injected into the `BuildGraph`. To inspect `traversable_dependency_specs`, one needs to full construct a `Target` object.

### Solution

As mentioned on #4420, the long term solution will be that the `BuildGraph` API goes away, and tasks directly request the products they need for various subjects. But for now, the only "dependency graph" that actually matters to frontend tasks is the graph defined by the `BuildGraph` API... and so we expose that by initializing a `BuildGraph` and then walking that to detect transitive changed targets.

### Result

Fixes #4420.