Fix resource loading issue in the python eval task. (#4452)
Turns out that `..` is not allowed in the path (which makes sense),
but this worked in practice when running from loose sources.

This change copies the relevant resource to the 'tasks2' dir, so that
the new task can load it from there.  It's a shame about the duplication,
but it's only temporary.