[engine] Use RuleGraph for task lookup instead of Tasks (#4371)
### Problem

With the RuleGraph we've built, we've got a lot of static info available to us about what the execution graph will contain. But, we're currently not using it to do task lookup.

### Solution

This patch changes the scheduler so that it uses the rule graph to do lookup instead of the Tasks collection.

### Result

As it is now it doesn't introduce a performance improvement, in my limited testing. But once we're using the static graph as the source for tasks to execute, there's lots of optimizations we can do.

### Problems / Follow on work

The implementation here is still pretty panicky. I think it would make sense to extract some Results or Options in a few more cases.