Add ParseContext singleton helper (#4466)
### Problem

Because no capability for parse-local state existed, `IntermediateTargetFactory` used class-local state to determine whether a singleton intermediate target had been created, rather than parse-local state.

The result was fragile... a subtle symptom (#4416) was that after a buildfile was invalidated in the daemon, the singleton targets it contained would not be recreated, causing them to go missing completely. A more obvious symptom was that `init.py` needed to reset the class-level state to ensure that intermediate targets could be recreated.

### Solution

Add support for parse-level state via a new method `create_object_if_not_exists` on `ParseContext`, and mark the already effectively-public methods of `ParseContext` as such.

As an orthogonal change, move from using a lock to guard the collection of created objects to using a thread-local collection.

### Result

Daemon runs that involve intermediate/'scoped'/'provided' targets that have been invalidated will succeed. Fixes #4416