Upgrade pants to current versions of pytest et al. (#4410)
Recent versions of pytest have a "file" attribute in the
junit xml they emit, which allowed us to simplify the logic
for mapping failed tests to targets.  But the new logic
no longer re-uses the junit xml parser helper we use on the
JVM side, as it doesn't support this 'file' attribute (it's
unclear if JUnit populates it) and modifying it to do so would
have made that helper more cumbersome than it's worth. Since
the new logic is so trivial, this isn't an issue.

Recent pytest also has a much simpler way of turning off
console reporting than the previous hack, now removed.

This change also ensures that the paths in coverage reports
are displayed relative to the buildroot.  This was the case
in the old pipeline, but not in the new.