Make pytest report sources paths relative to the buildroot. (#4472)
Previously the new PytestRun task reported paths relative to
the chroot, which was less readable for end users.

Note that this required writing a root conftest.py into the
chroot. Previously we wrote one (for reporting on sharding)
into its own module in a tmpdir, but now we require the
conftest.py hooks to be run when loading each file from the
chroot, which means it has to be in that chroot.

This makes things a little more complicated if the code
happens to have a root-level conftest.py already.  We take
some care to temporarily append to this conftest, and then
restore it to its original state.  We also take care to be
resilient to failures that somehow prevent the cleanup from
happening.