Get rid of the '2' registrations of the new python tasks.
That is, run2, test2 etc.

We no longer need them, and it's tricky to support them (e.g., they
may be stomping on the regular versions' products in weird ways).

Also removes some integration tests of the old python pipeline that
are now just invoking the new one, redundantly with the new one's
own integration tests.

It's not ideal not to have integration tests on the old code, but
since it's deprecated, and has been around forever, and still has
unittests, it should be OK. It doesn't seem worth futzing around
with config in those tests to make sure they invoke the old tasks.

I suspect that this will fix https://github.com/pantsbuild/pants/issues/4485