[engine] Use enum for RuleEdges keys, add factory for Selects w/o variants (#4461)
### Problem
selector paths between entries in the rule graph are not as well defined as they could be. Because of that, the `nodes.rs` usages of `RuleEdges` end up needing to do a lot of subject type filtering after use.

### Solution

This introduces a SelectKey enum that better clarifies the entry -> selector -> entries relationship. It represents the projections to other subjects as well as the nested selectors, which means we don't have to do subject filtering on the use side except in a couple cases.