Fix pytest fixture registration bug. (#4497)
This required changing how we hook into pytest to modify
how it displays paths.  The new hook is slightly more subtle,
but has the advantage of not having to be present during
test collection, which means we can put it in a tmpdir and
not in the source root, which allows us to get rid of a bunch
of complicated code dealing with an existing conftest.py
in the source root.