[engine] Better error messages for missing targets
Currently we always print out a trace when a BUILD file has missing targets.

This brings the v2 behavior more in line with v1 by adding special cases for resolve errors to ensure we produce an actionable error message.

It doesn't cover transitive references of missing targets yet. See #3912 for some of the possible cases