[engine] Better error messages for missing targets (#4509)
### Problem
There's two problems.
1. For some classes of missing targets, we always print a trace. V1 doesn't, it displays a short error message that's pretty clear.
2. For other error cases, we just say that the spec was missing without further information.

### Solution

This patch brings v2's behavior more in line with v1's for the above cases. It adds special handling for `ResolveError`s in those cases.

It does not attempt to improve trace generation generally, or to communicate where a non-existent address is referenced from if it has references to it. Those are follow one bits of work.

See #3912 for a number of related cases.