Add support for 'deployable_archives' for go and cpp rules. (#4518)
### Problem

Tasks need a canonical way to depend on executables.

### Solution

Add support for the `deployable_archives` product-type to `go` and `cpp` rules, which is already used by the `java` and `python` rules. The mapping stores the executable name keyed under target and path, like done in the `java`/`python` rules.  Fixes #4506.