Include transitive Resources targets in PrepareResources. (#4569)
### Problem

The `resources` arg on JVM and Python targets is now deprecated, but the previous behaviour of only collecting directly declared resources onto the runtime classpath remains. The result for the JVM is that resources are only collected if they are directly depended on by a `JvmTarget`, rather than if they are located anywhere in the transitive dependencies of that target.

While this behaviour has existed for a while, it was difficult to run into it when resources were directly declared as "owned" by a particular target.

### Solution

Transitively collect the relevant `Resources` targets below `JvmTarget` roots in `PrepareResources`.

Also, I've verified on a very large codebase that this did not result in any adverse effects on resource availability.

### Result

`Resources` act (more: see #4568) like other dependencies generally do. Fixes #4567