Specify a workunit for node.js test and run. (#4572)
### Problem

Not specifying the workunit for these tasks resulted in the binary name being used as the workunit name, which was very long (especially in CI environments) and not particularly readable. And because the workunit name ends up used as a filename, it was also possible to trip filename length limits this way.

### Solution

Align with the other node tasks in using the target address as the workunit name.

### Result

Output looks like
```
11:02:45 00:03     [node]
11:02:48 00:06       [examplebird/intake-service:node-lint]

                     > ...
```
rather than:
```
11:02:45 00:03     [node]
11:02:48 00:06       [/Users/stuhood/.cache/pants/bin/node/mac/10.11/v6.9.5/node/bin/npm]

                    > ...
```

And filename length limits aren't triggered in cases where the HOME directory path is very long.