Include API that will store target info in run_tracker (#4561)
### Problem
We would like to collect stats about the targets and tests being run so we would like to add a way to track this info. 

### Solution

The change included some refactoring to ease code re-use, a method to parse all the test information from the xml files, and two additional methods to collect target info and test info.

### Result

The changes shouldn't affect the end-user behavior. This change is for future development.