Fix export-classpaths exclude behavior
### Problem

Excluded jars were being included in the manifest jar created by export-classpath.

### Solution

Use the root targets as the starting point instead of the full closure when generating the classpath. When the whole closure was passed, it meant that even the jar_library targets were treated as root targets for the purposes of applying excludes, which meant that the least constrained excludes were always used.

### Result

Excludes are properly respected by classpaths from export-classpath