Switch to a conditional deprecation for the list-targets behaviour change. (#4600)
### Problem

The behaviour of scanning the entire repo when `./pants list` is called with no arguments is going away. We had a warning in place, but because `./pants list` is a console task, all logging was silenced.

### Solution

The warning should actually be a deprecation anyway, so switch to deprecated conditional. Additionally, make an executive decision as to the final behaviour: `./pants list` doing nothing at all would be more confusing than just throwing an error, so let's commit to throwing an error in the future.

### Result

Fixes #4597.