Fix broken export-classpath
### Problem

I'd assumed that ClassPathUtils.classpath would take just the root targets, and that it didn't apply excludes correctly when the closure was passed. That was incorrect.

Doing that would cause export-classpath to return an invalid classpath

### Solution

Revert that change, and update the regression test to be more comprehensive.

### Result

export-classpath works correctly again, and the regression test covers the excludes case more effectively