Improve Snapshot determinism (#4614)
### Problem

In #4610, we determined that the file mode is taken into account while creating Snapshots. But since git does not enforce file modes with the exception of the `execute` bit, the `umask` on various machines ends up influencing the Snapshot fingerprints.

### Solution

Implement a change to the `Deterministic` append mode upstream in `tar-rs`, and preemptively include it. The append mode will either be accepted as-is, or as a new append mode, in which case I'll update this review.

### Result

File mode is not included in the fingerprint, and so cross-platform/machine fingerprints are stable. Tested with checkouts with differing `umask` values across OSX and Linux. Fixes #4610.